+++ R/chart_labels.R
@@ -10,13 +10,13 @@
 #' mylc <- chart_labels(mylc, title = "my title", xlab = "my x label",
 #'   ylab = "my y label")
 chart_labels <- function( x, title = NULL, xlab = NULL, ylab = NULL){
-  if( !is.null(title) ) x$labels[["title"]] <- htmlEscape(title)
+  if( !is.null(title) ) x$labels[["title"]] <- title
   else x$labels[["title"]] <- NULL
 
-  if( !is.null(xlab) ) x$labels[["x"]] <- htmlEscape(xlab)
+  if( !is.null(xlab) ) x$labels[["x"]] <- xlab
   else x$labels[["x"]] <- NULL
 
-  if( !is.null(ylab) ) x$labels[["y"]] <- htmlEscape(ylab)
+  if( !is.null(ylab) ) x$labels[["y"]] <- ylab
   else x$labels[["y"]] <- NULL
   x
 }