+++ NEWS.md
@@ -5,6 +5,10 @@
 * Support for openxlsx2 by Jan Marvin Garbuszus
 * option to add table of data below corresponding levels by Marlon Molina
 
+## Issues
+
+* fix issue with % in labels of the graphic
+
 # mschart 0.3.1
 
 ## New features